rename muxed_conn
